regex-automata: fix deserialization of dense DFAs
This fixes an issue when deserializing dense DFAs where it could fail if
the bytes weren't properly aligned. We fix this by applying a clever
trick to force the compiler to align the result of `include_bytes!` to
the amount we want.

Much thanks to @ExpHP for sharing this hack:
https://users.rust-lang.org/t/can-i-conveniently-compile-bytes-into-a-rust-program-with-a-specific-alignment/24049/2