errors: improve failure modes when emoji-data.txt is missing
This commit improves the overall failure modes when I/O errors occur
while reading UCD data. In particular, the emoji-data.txt file is not
parse of the main UCD download, so we permit reading it to fail if it's
an I/O error.

Error messages should now include file paths.

Fixes #9