ucd-util: fix treatment of media hyphens
The implementation is slightly off. A medial hyphen only occurs when it
is between two letters, but we were implementing something subtly
different: a hyphen *wasn't* medial if it was surrounded by spaces. But
this neglects underscores and digits.

We also turn off the ellipsis deprecation warning for 1.24.1 builds. That
MSRV is needed by regex-syntax.