cli: add new --flat-table flag to case-related commands
This permits writing tables using '[char; 3]' instead of '&[char]'. The
latter is a fat point taking up 16 bytes by itself, where as the former
is an inlined sequence of 3 codepoints, using only 12 bytes. Since the
new representation drops the length encoded in '&[char]', we use the NUL
codepoint as a sentinel.

See #33 for a more in depth description with additional upsides of this
new mode.

Closes #33, Closes #34