cli: add --combined flag
This adds a new --combined flag to several commands. It permits
grouping all codepoints into a single table. If only a subset of
properties are being generated, and mere existence in one of the tables
is the only thing required, then this can be much more efficient since
one only needs to consult a single table.

PR #37