cli: remove the 'dfa' and 'regex' sub-commands
These have been moved to regex-cli and now use regex-automata 0.3:
https://github.com/rust-lang/regex/blob/master/regex-cli/README.md#example-serialize-a-dfa

This also breaks the cyclic dependency where updating to a new Unicode
version for bstr required the following:

* Run ucd-generate to update regex-syntax tables.
* Publish new regex-syntax.
* Update ucd-generate lockfile to bring in new regex-syntax.
* Build new ucd-generate binary.
* Run ucd-generate to update bstr regexes.

Namely, that last step requires updating regex-syntax in order to
propagate the Unicode updates into the regex engine.

The new process is:

* Run ucd-generate to update regex-syntax tables.
* Build regex-cli (also in the regex crate repo).
* Run regex-cli to update bstr regexes.

So now we don't have to do this weird dance where we loop back around to
build a new version of ucd-generate.

ucd-generate does still depend on `regex` at the moment via
`ucd-parse`, but this doesn't need updating when a new version of
Unicode comes out. Still, I'm going to explore breaking that dependency
as well via `regex-lite`.

ucd-generate also still depends on `ucd-util` which also has Unicode
data embedded into it. I'm going to look into fixing that by requiring
the caller to pass in the data tables.

Fixes #11